feat(core-support): implements enum style proc macro
