+++ crates/core/src/value/mod.rs
@@ -1,10 +1,12 @@
 //! Elvis values
+mod border;
 mod color;
 pub mod layouts;
 mod typo;
 mod unit;
 
 pub use {
+    border::BorderStyle,
     color::Colors,
     typo::{FontFamily, FontStyle},
     unit::Unit,
