+++ src/widgets/input.rs
@@ -1,0 +1,2 @@
+/// Interactive controls for Form
+pub struct Input;
