+++ src/widgets/mod.rs
@@ -1,6 +1,7 @@
 //! Evlis common widgets
 mod image;
+mod input;
 pub mod layouts;
 mod text;
 
-pub use {image::Image, text::Text};
+pub use {image::Image, input::Input, text::Text};
